Fix some build errors, but definitely not all of them.
* .settings/org.eclipse.jdt.core.prefs: To fix some build errors
  related to Java 1.7+ compliance (like not being able to switch on
  strings), I made the Eclipse IDE update the compliance settings at
  the project level.  This file was the result, so I versioned it.

* lib/: New folder.  Contains everything we needed to get the
  2018-testing branch of the 2017bot working from this weekend, except
  for CTRE_Phoenix-sources.jar, which I'm not convinced we need.

At runtime, we still seem to be putting the original WPILib files in
the classpath rather than the copies in our ./lib/ folder, but that's
fine by me as long as 2018bot builds on both Windows and Raspbian.
The lib/ folder is just a fallback to avoid hard-coding classpaths
anyway.

Currently broken on DifferentialDrive.  Not sure why, looking into it.