We edited the classpath in order to ensure that we were able to access all parts of the WPILib.jar consistently.
To do so, we first deleted the existing files. We then took our .jar
files out of build/jars and put them into lib/. This was done because
while building, build/jars is deleted and reformed. This ensured that
the faulty libraries were continually being used. Once the the jars
were moved and edited, we had to ensure that the build.xml was
referring to the correct folders. In build.properties, we set the
userLibs.dir to refer to the libs file in the workspace rather than
defaulting to a preset, device-specific location. It now refers to the
project directory's lib folder,${basedir}/lib.