Fix RobotMap and OI to better reflect port assignments.
In the past, our buttons and motor variable assignments were named very similarly. This led to confusion on whether or not OI was for buttons or motors, and led to reduntant or erroneous code. These variables have been now named FOO_BUTTON and FOO_PORT respectively.

Additionally, the references to such variables in the subclasses have been fixed appropriately.