Add climber subclass.
The climber subclass is very similar to our other subclasses,
comprising of two motors, keyed to Sparks (rather than our standard
Talons), and two buttons. When 'climbUp' is pressed (J on the
keyboard, or right 7 on joysticks), the two motors go full power
forward, extending the climber. When 'climbDown' is pressed (K on the
keyboard, or left 7 on joysticks), the two motors go full power
backward, retracting the climber. This class is highly subject to
change, as we still don't have a very clear design on the climber.