Figured out how to recursively explode selected JAR files.
It wasn't easy.

This is only half of the problem, though; while networktables-input.jar now
contains all the class files from all the JARs in the classpath, it still
doesn't run because it needs to find the ntcore native library in
java.library.path (analogous to $LD_LIBRARY_PATH).

I tried:

  LD_LIBRARY_PATH=~/wpilib/cpp/current/lib/ java -jar ./networktables-input.jar

It _almost_ worked, but the libntcore.so in the ~/wpilib/cpp/current/lib/
folder is an ARM build and not native to my platform (MageiaLinux 6.)  Hmm.