Create a distribution zip file that can run on the local platform.
Well, we've gotten to this point: the JAR can be built on a regular basis,
we zip it up with the right native library next to it. and there's even a
shell script to run it all.

The problem?  It _still_ fails.

  $ source ./networktables-input.sh
  > Exception in Application start method
  > java.lang.reflect.InvocationTargetException
  > [snip]
  > Caused by: java.lang.RuntimeException: Exception in Application start method
  > [snip]
  > Caused by: java.lang.UnsatisfiedLinkError: edu.wpi.first.networktables.NetworkTablesJNI.getDefaultInstance()I
  > [snip]
  > Exception running application main.NetworkTablesInput

Not sure what's causing this one yet, but I'm committing anyway.