Add some skeleton features to Turn.java
The Turn.java class is meant to handle turning from the robot's current alignment to the alignment needed to get to the next point. It takes in gyro data and compares that with the angle between our heading and the known necessary heading. Once the difference between the angles is 0 (or very close to 0), it stops and is ready to move on.