Modify Vector2 and Point
The Point class is much like the Vector2 class in that it stores a set of coordinates. Points can be converted to vectors and vectors can be converted to points. In the Turn command, speeds have been set. If the angle is greater than 45, the robot turns at full speed. If the angle is greater than 15, but less than 45, the robot turns at half speed. If the angle is greater than 0, but less than 15, the robot turns at half speed.