Add Go function for movement to points directly ahead or behind
The go function utilizes the accelerometer to get data about our position, which is calculated with a modified form of the kinematics. This iteratively uses these equations to change our current position and compare with our wanted position. With this data, it moves closer to the point until it is within an acceptable range. Comments are still needed.