Fix Intake and Launcher
Due to changes to the intake and launcher subsystems, some changes were made. A DoubleSolenoid was added to intake, as was a set of wheels. When the intake system lifts up, it will draw the cube in before lifting it up with the tank treads. When the intake system lowers down, the tank treads deposit the cube onto the arm before it spits it out. This will be used for scoring on the switch.