Key arm to right throttle
Instead of using a button to toggle the arm's position, it has instead been keyed to the throttle on the right stick. If the throttle reads a value greater than 0 (the throttle's angle relative to the ground is greater than 45 degrees), it will raise the arm. If it reads a value less than 0, then the arm will lower.