Fixed the octant and directions problem.
The angleBetween() function (which I implemented incorrectly, by the way) was
not the appropriate tool for the job of calculating octants.  The underlying
for product can't distinguish between octants above the X axis and octants
below.  The real function I needed was atan2(), which was expressly written
for this specific purpose.

A little displacement to move the range from [-π, π] to [0, 2π], a little
rounding to place the arrow character in the right place, a few more test
vectors to prove that the whole thing worked, and all was well.