Oh, so _that_ was why it wasn't moving!
Point.add() really ought to be called Point.sum(), since it does not change
the Point.

That being said, at least now the robot is moving, though it's not moving in
the right direction 6/8 of the time.  It's still better than before!

Increased the framerate and the movement speed in order to make the movement
errors easier to spot.