Intermediate checkpoint for the fake tank drive.
I think I have a mathematical model now for how the tank drive should
operate.  The key, as explained in the giant comment for
FakeRobotModel.FakeTankDrive.tankDrive(), is to model the robot as a line
segment with separate imparted rotational velocities for its left and right
sides.

Some infrastructure was needed, like vector and point rotation
functions--none of which are tested.