It turns out that our map center really is (0, 0).
When I made my mrior changes, I didn't pay attention to the coordinates of the
waypoints.  Changing this back, everything now works fine.

There's one remaining bug.  You'll note that the robot does not go all the way
to the left or all the way to the right--that's because we've been
"overdrawing" the map this whole time, painting the entire area of the screen
even though only a portion of it is actually used.  I'm working on fixing that
now.