Implemented a fake gyro class.
Unlike the FakeEncoder, this one is so untested that I haven't even called it
yet.  But it operates the same way as a FakeEncoder, and should be just as
useful (or useless.)

I didn't need to write a new interface for this since WPILibJ already has a
convenient edu.wpi.first.wpilibj.interfaces.Gyro interface that does exactly
what I need.

You're going to need to do a little vector math to get the value that needs to
be passed into updateAngle().  Maybe I should make that easier for the caller
with an updateHeading(Vector2) function?