Added the ability to highlight certain nodes.
We are trying to write a new simulation that tracks the robot as it
uses the GoEncoder and Turn Commands to navigate from waypoint to
waypoint.  Along the way, we want to highlight the next waypoint we're
traveling to and the goal waypoint (there's always a goal.)

The code in this commit adds the infrastructure for this.  We can
highlight a node, unhighlight it, and we draw the ones that are
highlighted.