Add FakeScheduler
The FakeScheduler class serves a function very similar to the WPILib
Scheduler class. Commands can be added to the list with the add()
method, and the run() method will execute the next command in the list.
Currently, the run() method will not work, as it simply executes
commands without regard for whether or not the previous command
functions.