I think I've fixed the TurnCommand turning problem.  Still testing, though.
The keys were threefold:

1. A better formula for calculating the minimum turning angle between
   two angles, source and dest:

   NORM(dest - source),       if NORM(dest - source) <= 180°
   NORM(dest - source) - 360, otherwise

   Where NORM(x) is a version of x modulo 360 that works on negative
   numbers.
2. The tank drive was being instructed to turn in the wrong direction.
3. We were calculating the estimated current heading incorrectly
   (gyro.getAngle() returns degrees, not radians.)

With those changes in place, the TurnCommand seems to turn in the
right direction, albeit slowly, and then its isFinished() returns
true.  Unfortunately, the GoEncoder seems to return isFinished()
immediately, so the robot does not move.  Once I've cleaned up the
TurnCommand, that's next on the agenda.