Fix TurnCommand
Finally, success! To a point at least. We now have a functioning TurnCommand, which will successfully turn the appropriate distance for various points. I have tested at various angles and orientations, and have gotten results accurate within .9 degrees. We are still unable to move from point to point. I believe that the issues are somewhere within the FollowPath, though it may lie with GoEncoder. To ensure there are not issues with the Interfaces and Wrappers we created, I have changed the mentions of the EncoderWrapper/Interface to the Encoder class. In order to build, I had to comment out the WaypointSimulator, though I have left it unaltered in this commit. If I am able to ensure that our autonomous works, I may remove the simulator from this branch.