Work on autonomous code, robot now can turn and STARTS go command. Does not currently end go command.
We learned that CommandGroup does not allow the start function, so we changed the start function and now the go command starts.
However an issue with the encoder has caused the distance to supposedly remain at zero the entire time, and thus the go command
never ends. But for the time being autonomous does start.

(cherry picked from commit fe53aeaef36628ce5693ada7cb90358b70418391)