Added six intermediate starting positions so that the robot has space to turn.
The robot (for which we assume a "diameter" of 36 inches) will generally start the match flush to the edge of the alliance station.  From theere, we're supposed to turn, but turning while you're against a wall will push against said wall, leading to inaccuracies that may interfere with our dead reckoning capabilities.

Better to add waypoints just a little forward of the starting positions, so that we can drive forward before we start turning.

NOTE: We don't have the command changes to drive forward yet.