Add slowdown method to intake
The intake system has been operating at full power, which can be innacurate during the acquisition of cubes for powerup and placement in the exchange. To rectify this, we have added a function that will half the intake's speed. This will allow us to be more precise around the exchange, a feature that has commonly been requested.