This is worthy of a commit, at least.
1) The problem I mentioned yesterday with the constantly-firing Spark
 motor controllers was caused by the team accidentally wiring the Sparks
 *backward*...and me not noticing!

2) Our potentiometer is flaky.  It actually produces output that varies
 between 0 and 0.27 rather than between 0 and 1.  I can work around
this in software, but that requires linear inteprolation at several stages of
the process.

* First, I measured the output of the pot with a SCALE_FACTOR of 1.0
 and an OFFSET of 0.0..  Dividing the actual pot output by the difference
between these two values (i.e., 0.27) will produce a numebr that varies
between 0 and 1.

* Scaling the SCALE_FACTOR itself by the difference between my desired
maximum (300 degrees = 10:00) and minimum (60 degrees = 2:00)
produced an actual, working SCALE_FACTOR to pass into the constructor
for the pot.

* So now the pot varies correctly between MIN_OUTPUT_POT_VALUE
and MAX_OUTPUT_POT_VALUE.  The other change I made ensures that
we can get a "u" value between 0 and 1 from that.  We then use that
value (which I have also confirmed to be calculated correctly) to
determine the motor power.

I'm falling short of my expectations on that last bit, but the rest is working alright.
solidly.