Rewrite the test bed code to use a better potentiometer.
We replaced the first potentiometer with a couple of nicer pots with
kilo-ohm resistances (rather than the megaohm resistances the old pot
apparently had.)  The response curve on the pots is linear, so when it
outputs 0.5, it really is at the halfway mark.

Using this, I rewrote the code and it works perfectly -- the pot can be
used to turn the motor counterclockwise or clockwise with fairly fine
adjustments.  Not all of our motors will be CIMs, so
RobotMap.MIN_MOTOR_POWER will probably have to change at some point.

But analog turning-knob controls, while easy, are not what we're here for.
The next step is harder: as the pot rotates, the motor must rotate as
well--their motions need to be linked.  Only _then_ can I start work on
a PID subsystem that moves the arm motor to set positions.

Really, if you think about it, we're making a servo by hand here.  The
firste step is finishing the hardware.