Elbow controls now use L/R button 5 to change position, and R button 6 to commit PID movement.
We don't have something similar for the wrist yet; that's still manually
controlled.  Furthermore, the DESIRED_POSITIONS[] array contains made-up
values.  During the lead-up to the competition, I fully expect that the
students will measure better values for the elbow position and alter the
number of entries and/or the values for DESIRED_POSITIONS[].