Added sneaking code!
First, we have a "sneak" object that maintains state -- a double value
that's either 1.0 or not 1.0.  We cna get this value at any time.
Robot.java has a handleSneak() function that changes the drive and wrist
sneak values (they're separate), and changes them back.  The wrist and
drive use these sneak values as force multipliers.

The verdict: Drive sneak works great!  The wrist sneak is a little too
slow and timid, but it's at least usable.  That's a worthy checkpoint
for committing.